# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/student/profile": {
      "put": {
        "operationId": "StudentController_upsertProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertStudentProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile upserted successfully."
          }
        },
        "summary": "Upsert student profile",
        "tags": [
          "Student"
        ]
      },
      "get": {
        "operationId": "StudentController_getMyProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully."
          }
        },
        "summary": "Get student profile for frontend store",
        "tags": [
          "Student"
        ]
      }
    },
...
    "/api/student/application/{resourceId}": {
      "delete": {
        "operationId": "ApplicationController_cancelApplication",
        "parameters": [
          {
            "name": "resourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Application canceled successfully."
          }
        },
        "summary": "Cancel a student application",
        "tags": [
          "Application"
        ]
      }
    },
    "/api/student/application/create": {
      "post": {
        "operationId": "ApplicationController_createApplication",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application created successfully."
          }
        },
        "summary": "Create a new application",
        "tags": [
          "Application"
        ]
      }
    },
...
  },
  "info": {
    "title": "API docs",
    "description": "My NestJS API manual",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UpsertStudentProfileDto": {
        "type": "object",
        "properties": {}
      },
      "CreateApplicationDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}